ISSD-3030 Improve performance of filter component Selects
